Allow custom string delimiters
Some projects prefer `'` to delimite JavaScript strings, others prefer
`"`. The mocha code in a project should obviously follow suit, but right
now snippets are hard-coded to use `"`. This changes the default to `'`
since it seems to be gaining in popularity, but it also adds the custom
variable `mocha-snippets-string-delimiter` which will be used for all of
the strings in its templates.

To customize the string delimiter for a single project to be the
backtick for example, you could do so by putting the following pair in
your `.dir-locals.el` file.

```elisp
((js2-mode . ((mocha-snippets-string-delimiter . "`"))))
```